requirements.txt
starkbank-ecdsa==0.1.4
bitcoinlib